+++ tests/test_iqm_job.py
@@ -16,9 +16,18 @@
 """
 import uuid
 
-from iqm_client import HeraldingMode, Instruction, IQMClient, RunResult, RunStatus, SingleQubitMapping, Status
+from iqm_client import (
+    HeraldingMode,
+    Instruction,
+    IQMClient,
+    JobAbortionError,
+    RunResult,
+    RunStatus,
+    SingleQubitMapping,
+    Status,
+)
 import mockito
-from mockito import mock, when
+from mockito import mock, unstub, verify, when
 import pytest
 from qiskit import QuantumCircuit
 from qiskit.providers import JobStatus
@@ -63,14 +72,46 @@ def iqm_metadata():
     }
 
 
+@pytest.fixture()
+def iqm_metadata_with_timestamps():
+    measurement = Instruction(name='measurement', implementation=None, qubits=('0',), args={'key': 'm1'})
+    return {
+        'calibration_set_id': 'df124054-f6d8-41f9-b880-8487f90018f9',
+        'request': {
+            'shots': 4,
+            'circuits': [{'name': 'circuit_1', 'instructions': (measurement,), 'metadata': {'a': 'b'}}],
+            'calibration_set_id': 'df124054-f6d8-41f9-b880-8487f90018f9',
+            'qubit_mapping': [
+                SingleQubitMapping(logical_name='0', physical_name='QB1'),
+                SingleQubitMapping(logical_name='1', physical_name='QB2'),
+            ],
+        },
+        'timestamps': {
+            'job_start': 0.0,
+            'job_end': 2.3,
+        },
+    }
+
+
 def test_submit_raises(job):
     with pytest.raises(NotImplementedError, match='You should never have to submit jobs by calling this method.'):
         job.submit()
 
 
-def test_cancel_raises(job):
-    with pytest.raises(NotImplementedError, match='Canceling jobs is currently not supported.'):
-        job.cancel()
+def test_cancel_successful(job, recwarn):
+    when(job._client).abort_job(uuid.UUID(job.job_id())).thenReturn(None)
+    assert job.cancel() is True
+    assert len(recwarn) == 0
+    verify(job._client, times=1).abort_job(uuid.UUID(job.job_id()))
+    unstub()
+
+
+def test_cancel_failed(job):
+    when(job._client).abort_job(uuid.UUID(job.job_id())).thenRaise(JobAbortionError)
+    with pytest.warns(UserWarning, match='Failed to cancel job'):
+        assert job.cancel() is False
+    verify(job._client, times=1).abort_job(uuid.UUID(job.job_id()))
+    unstub()
 
 
 def test_status_for_ready_result(job):
@@ -88,15 +129,18 @@ def test_status_done(job, iqm_metadata):
     assert job._result is None
 
 
-@pytest.mark.parametrize('run_status', [Status.PENDING_COMPILATION, Status.PENDING_EXECUTION])
-def test_status_running(job, run_status):
+@pytest.mark.parametrize(
+    'run_status,job_status',
+    [
+        (Status.PENDING_COMPILATION, JobStatus.QUEUED),
+        (Status.PENDING_EXECUTION, JobStatus.RUNNING),
+        (Status.FAILED, JobStatus.ERROR),
+        (Status.ABORTED, JobStatus.CANCELLED),
+    ],
+)
+def test_other_job_statuses(job, run_status: Status, job_status: JobStatus):
     when(job._client).get_run_status(uuid.UUID(job.job_id())).thenReturn(RunStatus(status=run_status))
-    assert job.status() == JobStatus.RUNNING
-
-
-def test_status_fail(job):
-    when(job._client).get_run_status(uuid.UUID(job.job_id())).thenReturn(RunStatus(status=Status.FAILED))
-    assert job.status() == JobStatus.ERROR
+    assert job.status() == job_status
 
 
 def test_result(job, iqm_result_two_registers, iqm_metadata):
@@ -179,3 +223,19 @@ def test_result_multiple_circuits(job, iqm_result_two_registers):
         assert r.calibration_set_id == uuid.UUID('9d75904b-0c93-461f-b1dc-bd200cfad1f1')
         assert r.data.metadata == {'a': i}
     assert result.request.qubit_mapping == iqm_metadata_multiple_circuits['request']['qubit_mapping']
+
+
+def test_result_with_timestamps(job, iqm_result_two_registers, iqm_metadata_with_timestamps):
+    client_result = RunResult(
+        status=Status.READY,
+        measurements=[iqm_result_two_registers],
+        metadata=iqm_metadata_with_timestamps,
+    )
+    when(job._client).wait_for_results(uuid.UUID(job.job_id())).thenReturn(client_result)
+
+    assert job.metadata.get('timestamps') is None
+    result = job.result()
+    assert 'timestamps' in job.metadata
+    assert job.metadata.pop('timestamps') == iqm_metadata_with_timestamps.get('timestamps')
+    assert 'timestamps' in result._metadata
+    assert result.timestamps == iqm_metadata_with_timestamps.get('timestamps')